inliner aggressiveness as i64

+++ noir/noir-repo/compiler/noirc_driver/src/lib.rs
@@ -128,8 +128,8 @@ pub struct CompileOptions {
     /// Setting to decide on an inlining strategy for brillig functions.
     /// A more aggressive inliner should generate larger programs but more optimized
     /// A less aggressive inliner should generate smaller programs
-    #[arg(long, hide = true, allow_hyphen_values = true, default_value_t = f64::INFINITY)]
-    pub inliner_aggressiveness: f64,
+    #[arg(long, hide = true, allow_hyphen_values = true, default_value_t = i64::MAX)]
+    pub inliner_aggressiveness: i64,
 }
 
 pub fn parse_expression_width(input: &str) -> Result<ExpressionWidth, std::io::Error> {
